Modified the code to match in db for displaying in UI.

+++ app/src/main/java/com/techradicle/expensetracker/domain/model/ocr/Items.kt
@@ -3,5 +3,5 @@ package com.techradicle.expensetracker.domain.model.ocr
 data class Items(
     val amount: Double,
     val description: String,
-    val qty: Int
+    val qty: Double
 )
